// packages/stencil-library/src/utils/component-props-types.ts
export type ExtendedColorType =
    | "primary"
    | "secondary"
    | "auto"
    | "inverse"
    | "light"
    | "white"
    | "black"
    | "success"
    | "warning"
    | "danger"
    | "info";
// ...
export type BorderRadiusType =
    | "none"
    | "default"
    | "xs"
    | "sm"
    | "md"
    | "lg"
    | "xl"
    | "2xl"
    | "3xl"
    | "full"
    | "circle";
// ...
export type OptionalAppearanceType =
    | "transparent"
    | "none"
    | "solid"
    | "outlined"
    | "mixed";
// ...
export type ExtendedSizeType =
    | "xs"
    | "sm"
    | "md"
    | "lg"
    | "xl";